impl generic taint generic over string and u64
