+++ python/src/utils.rs
@@ -11,6 +11,8 @@ use raphtory::core::time::Interval;
 use raphtory::db::view_api::time::WindowSet;
 use raphtory::db::view_api::TimeOps;
 use std::error::Error;
+use raphtory::core::vertex::InputVertex;
+use raphtory::core as dbc;
 
 /// Extract a `VertexRef` from a Python object.
 /// The object can be a `str`, `u64` or `PyVertex`.
@@ -112,5 +114,52 @@ impl TryFrom<IntervalBox> for Interval {
     type Error = ParseTimeError;
     fn try_from(value: IntervalBox) -> Result<Self, Self::Error> {
         value.interval
+    }
+}
+
+/// A trait for vertices that can be used as input for the graph.
+/// This allows us to add vertices with different types of ids, either strings or ints.
+#[derive(Clone, Debug)]
+pub struct InputVertexBox {
+    id: u64,
+    name_prop: Option<dbc::Prop>,
+}
+
+/// Implementation for vertices that can be used as input for the graph.
+/// This allows us to add vertices with different types of ids, either strings or ints.
+impl InputVertexBox {
+    pub(crate) fn new<T>(vertex: T) -> InputVertexBox
+        where
+            T: InputVertex,
+    {
+        InputVertexBox {
+            id: vertex.id(),
+            name_prop: vertex.name_prop(),
+        }
+    }
+}
+
+/// Implementation for vertices that can be used as input for the graph.
+/// This allows us to add vertices with different types of ids, either strings or ints.
+impl InputVertex for InputVertexBox {
+    /// Returns the id of the vertex.
+    fn id(&self) -> u64 {
+        self.id
+    }
+
+    /// Returns the name property of the vertex.
+    fn name_prop(&self) -> Option<dbc::Prop> {
+        self.name_prop.clone()
+    }
+}
+
+pub(crate) fn extract_input_vertex(id: &PyAny) -> PyResult<InputVertexBox> {
+    match id.extract::<String>() {
+        Ok(string) => Ok(InputVertexBox::new(string)),
+        Err(_) => {
+            let msg = "IDs need to be strings or an unsigned integers";
+            let number = id.extract::<u64>().map_err(|_| PyTypeError::new_err(msg))?;
+            Ok(InputVertexBox::new(number))
+        }
     }
 }
