+++ python/tests/test_graphdb.py
@@ -845,10 +845,10 @@ def test_generic_taint():
 
     actual = algorithms.generic_taint(g, 20, 11, [1, 2], [4, 5])
     expected = {
-        5: [(6, 13, 2)],
-        1: [(0, 11, 1)],
-        4: [(5, 12, 2)],
-        2: [(2, 12, 1), (0, 11, 2), (2, 11, 1)]
+        '1': [(0, 11, '1')],
+        '2': [(2, 12, '1'), (2, 11, '1'), (0, 11, '2')],
+        '4': [(5, 12, '2')],
+        '5': [(6, 13, '2')],
     }
 
     assert(actual == expected)
